Map filteren

+++ Code/WijkAgent2/WijkAgent2/Classes/Delict.cs
@@ -13,8 +13,8 @@ namespace Wijkagent2.Classes
         public string zip { get; set; }
         public string street { get; set; }
         public StringBuilder description { get; set; }
-        public decimal longitude { get; set; }
-        public decimal lat { get; set; }
+        public double longitude { get; set; }
+        public double lat { get; set; }
         public int status { get; set; }
         
         public string person { get; set; }
